+++ docs/SCRIPT_GUIDELINES.md
